Added files from local storage.
The class as of now works with bare minimum.
Should check up on the rules in the case of a tie or a split pot
There are too many bugs which have not yet been discovered. 
The first one is comparison in the case of a TWO_PAIR and any other hand. The valueInHand is not being updated.